refactor: split writings tasks
