+++ handlers/writings/edit_reply_task.go
@@ -1,0 +1,92 @@
+package writings
+
+import (
+	"database/sql"
+	"errors"
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/gorilla/mux"
+
+	"github.com/arran4/goa4web/core"
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
+	"github.com/arran4/goa4web/handlers"
+	"github.com/arran4/goa4web/internal/db"
+	notif "github.com/arran4/goa4web/internal/notifications"
+	"github.com/arran4/goa4web/internal/tasks"
+	"github.com/arran4/goa4web/workers/postcountworker"
+)
+
+// EditReplyTask updates an existing comment.
+type EditReplyTask struct{ tasks.TaskString }
+
+var editReplyTask = &EditReplyTask{TaskString: TaskEditReply}
+
+var _ tasks.Task = (*EditReplyTask)(nil)
+var _ notif.AdminEmailTemplateProvider = (*EditReplyTask)(nil)
+
+func (EditReplyTask) Action(w http.ResponseWriter, r *http.Request) any {
+	languageID, err := strconv.Atoi(r.PostFormValue("language"))
+	if err != nil {
+		return fmt.Errorf("language parse fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+	text := r.PostFormValue("replytext")
+
+	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+	vars := mux.Vars(r)
+	articleID, err := strconv.Atoi(vars["article"])
+	if err != nil {
+		return fmt.Errorf("article id parse fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+	commentID, err := strconv.Atoi(vars["comment"])
+	if err != nil {
+		return fmt.Errorf("comment id parse fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+
+	session, ok := core.GetSessionOrFail(w, r)
+	if !ok {
+		return handlers.SessionFetchFail{}
+	}
+	uid, _ := session.Values["UID"].(int32)
+
+	comment := r.Context().Value(consts.KeyComment).(*db.GetCommentByIdForUserRow)
+
+	thread, err := queries.GetThreadLastPosterAndPerms(r.Context(), db.GetThreadLastPosterAndPermsParams{
+		ViewerID:      uid,
+		ThreadID:      comment.ForumthreadID,
+		ViewerMatchID: sql.NullInt32{Int32: uid, Valid: uid != 0},
+	})
+	if err != nil && !errors.Is(err, sql.ErrNoRows) {
+		return fmt.Errorf("get thread fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+
+	if err = queries.UpdateComment(r.Context(), db.UpdateCommentParams{
+		Idcomments:         int32(commentID),
+		LanguageIdlanguage: int32(languageID),
+		Text:               sql.NullString{String: text, Valid: true},
+	}); err != nil {
+		return fmt.Errorf("update comment fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+
+	if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
+		if evt := cd.Event(); evt != nil {
+			if evt.Data == nil {
+				evt.Data = map[string]any{}
+			}
+			evt.Data[postcountworker.EventKey] = postcountworker.UpdateEventData{ThreadID: thread.Idforumthread, TopicID: thread.ForumtopicIdforumtopic}
+		}
+	}
+
+	return handlers.RedirectHandler(fmt.Sprintf("/writings/article/%d", articleID))
+}
+
+func (EditReplyTask) AdminEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("adminNotificationNewsCommentEditEmail")
+}
+
+func (EditReplyTask) AdminInternalNotificationTemplate() *string {
+	v := notif.NotificationTemplateFilenameGenerator("adminNotificationNewsCommentEditEmail")
+	return &v
+}
