+++ handlers/writings/writing_category_change_task.go
@@ -1,0 +1,52 @@
+package writings
+
+import (
+	"database/sql"
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
+	"github.com/arran4/goa4web/handlers"
+	"github.com/arran4/goa4web/internal/db"
+	"github.com/arran4/goa4web/internal/tasks"
+)
+
+// WritingCategoryChangeTask modifies a category.
+type WritingCategoryChangeTask struct{ tasks.TaskString }
+
+var writingCategoryChangeTask = &WritingCategoryChangeTask{TaskString: TaskWritingCategoryChange}
+
+var _ tasks.Task = (*WritingCategoryChangeTask)(nil)
+
+func (WritingCategoryChangeTask) Action(w http.ResponseWriter, r *http.Request) any {
+	name := r.PostFormValue("name")
+	desc := r.PostFormValue("desc")
+	wcid, err := strconv.Atoi(r.PostFormValue("wcid"))
+	if err != nil {
+		return fmt.Errorf("wcid parse fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+	cid, err := strconv.Atoi(r.PostFormValue("cid"))
+	if err != nil {
+		return fmt.Errorf("cid parse fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+
+	if err := queries.UpdateWritingCategory(r.Context(), db.UpdateWritingCategoryParams{
+		Title: sql.NullString{
+			Valid:  true,
+			String: name,
+		},
+		Description: sql.NullString{
+			Valid:  true,
+			String: desc,
+		},
+		Idwritingcategory: int32(cid),
+		WritingCategoryID: int32(wcid),
+	}); err != nil {
+		return fmt.Errorf("update writing category fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+
+	return nil
+}
